GH-1444: Observablility Documentation, Polish
Refactor for latest snapshots; add documentation generation.
Use NOOP registry and supplier for context.

Also fix class tangles.

Resolves https://github.com/spring-projects/spring-amqp/issues/1507

Disable auto doc generation and polish manually.
