+++ build.gradle
@@ -1,5 +1,6 @@
 buildscript {
 	ext.kotlinVersion = '1.7.0'
+	ext.isCI = System.getenv('GITHUB_ACTION') || System.getenv('bamboo_buildKey')
 	repositories {
 		mavenCentral()
 		gradlePluginPortal()
@@ -55,8 +56,9 @@ ext {
 	log4jVersion = '2.18.0'
 	logbackVersion = '1.2.3'
 	lz4Version = '1.8.0'
-	micrometerVersion = '1.10.0-M6'
-	micrometerTracingVersion = '1.0.0-M8'
+	micrometerDocsVersion = '1.0.0-SNAPSHOT'
+	micrometerVersion = '1.10.0-SNAPSHOT'
+	micrometerTracingVersion = '1.0.0-SNAPSHOT'
 	mockitoVersion = '4.8.0'
 	rabbitmqStreamVersion = '0.8.0'
 	rabbitmqVersion = project.hasProperty('rabbitmqVersion') ? project.rabbitmqVersion : '5.16.0'
@@ -221,7 +223,7 @@ subprojects { subproject ->
 	}
 
 	task updateCopyrights {
-		onlyIf { !System.getenv('GITHUB_ACTION') && !System.getenv('bamboo_buildKey') }
+		onlyIf { !isCI }
 		inputs.files(modifiedFiles.filter { f -> f.path.contains(subproject.name) })
 		outputs.dir('build/classes')
 
@@ -374,6 +376,10 @@ project('spring-amqp') {
 project('spring-rabbit') {
 	description = 'Spring RabbitMQ Support'
 
+	configurations {
+		adoc
+	}
+
 	dependencies {
 
 		api project(':spring-amqp')
@@ -411,7 +417,36 @@ project('spring-rabbit') {
 		testRuntimeOnly 'com.fasterxml.jackson.module:jackson-module-kotlin'
 		testRuntimeOnly ("junit:junit:$junit4Version") {
 			exclude group: 'org.hamcrest', module: 'hamcrest-core'
+
+		adoc "io.micrometer:micrometer-docs-generator-spans:$micrometerDocsVersion"
+		adoc "io.micrometer:micrometer-docs-generator-metrics:$micrometerDocsVersion"
+
 		}
+
+		def inputDir = file('src/main/java/org/springframework/amqp/rabbit/support/micrometer').absolutePath
+		def outputDir = rootProject.file('src/reference/asciidoc').absolutePath
+
+		task generateObservabilityMetricsDocs(type: JavaExec) {
+			onlyIf { !isCI }
+			mainClass = 'io.micrometer.docs.metrics.DocsFromSources'
+			inputs.dir(inputDir)
+			outputs.dir(outputDir)
+			classpath configurations.adoc
+			args inputDir, '.*', outputDir
+		}
+
+		task generateObservabilitySpansDocs(type: JavaExec) {
+			onlyIf { !isCI }
+			mainClass = 'io.micrometer.docs.spans.DocsFromSources'
+			inputs.dir(inputDir)
+			outputs.dir(outputDir)
+			classpath configurations.adoc
+			args inputDir, '.*', outputDir
+		}
+
+		// javadoc {
+		// 	finalizedBy generateObservabilityMetricsDocs, generateObservabilitySpansDocs
+		// }
 
 	}
