+++ src/reference/asciidoc/_conventions.adoc
@@ -1,0 +1,11 @@
+[[observability-conventions]]
+=== Observability - Conventions
+
+Below you can find a list of all `GlobalObservabilityConventions` and `ObservabilityConventions` declared by this project.
+
+.ObservationConvention implementations
+|===
+|ObservationConvention Class Name | Applicable ObservationContext Class Name
+|`DefaultRabbitListenerObservationConvention`|`RabbitMessageReceiverContext`
+|`DefaultRabbitTemplateObservationConvention`|`RabbitMessageSenderContext`
+|===
