+++ src/reference/asciidoc/_metrics.adoc
@@ -1,0 +1,44 @@
+[[observability-metrics]]
+=== Observability - Metrics
+
+Below you can find a list of all samples declared by this project.
+
+[[observability-metrics-listener-observation]]
+==== Listener Observation
+
+____
+Observation for Rabbit listeners.
+____
+
+**Metric name** `spring.rabbit.listener` (defined by convention class `RabbitListenerObservation$DefaultRabbitListenerObservationConvention`). **Type** `timer` and **base unit** `seconds`.
+
+Name of the enclosing class `RabbitListenerObservation`.
+
+IMPORTANT: All tags must be prefixed with `spring.rabbit.listener` prefix!
+
+.Low cardinality Keys
+[cols="a,a"]
+|===
+|Name | Description
+|`spring.rabbit.listener.id`|Listener id.
+|===
+
+[[observability-metrics-template-observation]]
+==== Template Observation
+
+____
+Observation for `RabbitTemplate` s.
+____
+
+**Metric name** `spring.rabbit.template` (defined by convention class `RabbitTemplateObservation$DefaultRabbitTemplateObservationConvention`). **Type** `timer` and **base unit** `seconds`.
+
+Name of the enclosing class `RabbitTemplateObservation`.
+
+IMPORTANT: All tags must be prefixed with `spring.rabbit.template` prefix!
+
+.Low cardinality Keys
+[cols="a,a"]
+|===
+|Name | Description
+|`spring.rabbit.template.name`|Bean name of the template.
+|===
